# Performance improvements to sqlparser and changed the interface to allow io.Reader